Initial commit of Boostnote
Summary: Initial commit of Boostnote. This initial commit is for version 0.11.4 on purpose as the latest version (0.11.6) has some electron interface issues that cause the markdown notes to not be previewed correctly. I will continue working on packaging the latest version of Boostnote but because of the enthusiasm around this package, I thought I should post this early. This resolves T3207.

Test Plan: Tested by creating & deleting markdown notes as well as creating new storage locations. Everything functioned as expected.

Reviewers: #triage_team, sunnyflunk

Reviewed By: #triage_team, sunnyflunk

Maniphest Tasks: T3207

Differential Revision: https://dev.solus-project.com/D3114