Update boostnote to 0.15.3
Summary:
Fixes T8963.

This is an update for the legacy version of boostnote. The newer version mentioned in the Update reuqest was Boostnote.next, which is incompatible with the legacy version of the app. Therefor it should not silently replace the legacy version, but should be added as a new package with "replaces" flag.

Test Plan: Created/edited/deleted Multiple notes and folders. Tried opening an old note storage creatd with the previous version.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Maniphest Tasks: T8963

Differential Revision: https://dev.getsol.us/D8977